requirements.txt
telethon
tweepy
google-generativeai
supabase
flask
python-dotenv
asyncio
requests